add: task #1 Let's get some images!
The description of the project contains some inspiration for the final
look of the project but we’ll have to download some images.

Head to unsplash and download 10 high resolution images that look as
close to the final product that you’re going to make. You will be using
these same high res images for a project on Responsive Design in the
future. Remember to also include the 3 images (the 2 logos and the
favicon) linked in the description of the project.

The images should all be representative of category they belong to.
Images in the work category should be closely related to work.