add: task #7. Make the font size responsive
We have multiple ways to make the typography responsive. The basic
way would be to create multiple media queries and set a different
font-size. But because we are using REM that are based on 62.5%
(defined in the html selector). Changing that value would change
proportionally all font sizes.

In your CSS file at the end of the /* Base section

Create a new media query for max-width: 480px
Target the html element
Property: font-size, Value: 57%
Create a new media query for min-width: 481px and max-width: 767px
Target the html element
Property: font-size, Value 60%
This is a simple way to achieve responsive typography. More complex
options can also be used to have a more granular control over the
font sizes.